Type_Pipeline-Agent/requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-deepseek>=0.1.0
langgraph>=0.0.10
langchain-tavily>=0.1.0
